:construction: chore: Header component updated
